python_services/requirements.txt
# Core ML + service dependencies will be pinned when implementations land.
# Add CUDA-specific wheels per environment when wiring Whisper/pyannote.

fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
torch>=2.2.0
whisper-openai>=1.0.0
faster-whisper>=1.0.0
pyannote.audio>=3.0.0
vosk>=0.3.45
coqui-tts>=0.22.0
google-cloud-texttospeech>=2.16.3
azure-cognitiveservices-speech>=1.37.0
protobuf>=4.25.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
grpcio>=1.62.0
pytest>=8.0.0
numpy>=1.24.0
scipy>=1.10.0
openai>=1.0.0
soundfile>=0.12.0
onnxruntime>=1.16.0
